python-bot/requirements.txt
aiosqlite==0.20.0
aiohttp==3.9.5
aiogram==3.13.1
asyncpg==0.29.0
httpx==0.27.0
python-dotenv==1.0.1

